refactor(doctor): extract default account warning helpers
