+++ src/commands/doctor-config-flow.missing-explicit-default-account.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import type { OpenClawConfig } from "../config/config.js";
-import { collectMissingExplicitDefaultAccountWarnings } from "./doctor-config-flow.js";
+import { collectMissingExplicitDefaultAccountWarnings } from "./doctor-config-default-account-warnings.js";
 
 describe("collectMissingExplicitDefaultAccountWarnings", () => {
   it("warns when multiple named accounts are configured without default selection", () => {
